Actually raise a TooManyKeysError for more than 6 keys in trigger_keys
The hardware only supports up to six keys at a time, and this function should raise an exception when more keys are passed. However, the press_keys list was first reduced to six items and then checked, meaning it would never raise an exception and still silently ignored keys. Also added a unit test to verify this behavior.

Related to #12.