Raise exception when something fails
When the hardware sends an unexpected response, raise an exception.

Especially for get_parameters it didn't make sense to return an empty string, since this normally returns a bytes-like object. Exceptions are meant for this and make it easier to handle errors without the need of checking return values. Also, it makes it possible to distinguish between an error and sucessfully retrieving an empty string.