Add Configuration class
To make it easier to view and modify the various configuration fields. Configuration is a bytes-like object to remain backward compatible, but it has properties that do the packing and unpacking from the configuration byte string. This makes it easier to view and change the various configuration parameters.

The result of `get_parameters` now has more fields to easily access the configuration:

```
curconf = config.get_parameters(ser)
print(f"Current baud rate is {curconf.serial_comm_baud_rate}")
```

Currently this is missing the return characters configuration and the filter characters configuration. Also, it does not modify the checksum field automatically when changing the configuration fields. There is not a function yet to send a modified configuration back to the hardware.